채팅 파싱 구조 변경
- 한글의 정상적인 데이터 이동을 위한 파싱 코드 변경

+++ src/Server.h
@@ -102,7 +102,7 @@ public:
     void ParseAndWork(ClientInfo *client, Buffer *buf);
 
     /*방 생성*/
-    Room *GenerateRoom(char *title, int title_length);
+    Room *GenerateRoom(char *title, unsigned short  title_length);
 
     void DeleteRoom(ClientInfo *room);
